backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-community
langchain-text-splitters
langgraph
langchain-google-genai>=4.0.0
langchain-anthropic
langchain-openai
langsmith
chromadb
httpx
python-dotenv
pydantic